refactor(project): rename inventory_service to inventoryService
