+++ services/inventory/src/main/java/com/petek/inventoryService/repository/ProductRepository.java
@@ -1,9 +1,9 @@
-package com.petek.inventory_service.repository;
+package com.petek.inventoryService.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-import com.petek.inventory_service.entity.Product;
+import com.petek.inventoryService.entity.Product;
 
 public interface ProductRepository extends JpaRepository<Product, Long>, JpaSpecificationExecutor<Product> {
