+++ services/inventory/src/test/java/com/petek/inventoryService/InventoryServiceApplicationTests.java
@@ -1,4 +1,4 @@
-package com.petek.inventory_service;
+package com.petek.inventoryService;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
